Route form elements testing to distinct page
At branch creation, testing of basic reusable form components was in App.tsx. Testing has since been moved to a dedicated page, at path /FormElementsPage.
